Move UI popup menu functions to separate `ui_popups.cpp`
